refactored plan-resolve and cpp cli to use cpp module with a sane interface
The problem with the old cpp_opts module was that it violated
separation of concerns in a very ugly way:
* Instead of just handling cpp options, it also handled long options,
which complicated the interface, suggesting getopt like responsibility
while behaving unlike getopt.

* It prepared command suited specifically for execution by
util.system_pipe (which is really Popen in disguise), which
is unrelated to it. This creates an invisible dependency between
what parse returns and how it is supposed to be used. This relationship
is not explained, or documented, and it doesn't really make any sense either.

* The general problem with code like this is that it makes very specific
assumptions about who is using it and how, breaking the principle of
separation of concerns and making the code much more difficult to follow.

Each piece of code should really strive to have an interface that makes
logical sense without relation to the rest of the program.
*Especially* if you are creating a public interface such as a module.