Remove isolinux from run-genisoimage and replace with grub only approach.
The isohybrid approach, using isohybrid-gpt-basdat xorriso option results
in problems booting some laptops, as they seem not to recognize the efi partition.
We replace that approach with a grub only iso which uses a gpt partition
table in addition to a mbr (basically identical approach as grub-mkrescue does).
This has the advantage that isolinux is not needed anymore and still
the generated images are bootable on BIOS and UEFI systems.