Rebuild steps even with stamp
fab uses stamps to check if each deck is built or not. This is required
as existing deck's could of failed and still exist. However when a stamp
exists and the deck does not, it still thinks it does not need to
rebuild.

By making each stage dependent on both the stamp AND deck of the
previous step, stages will be rebuilt in this condition without having
to manually remove the stamp.